+++ blockchain/chainio.go
@@ -1272,12 +1272,12 @@ func (b *BlockChain) createChainState() error {
 	// Create a new node from the genesis block and set it as the best node.
 	genesisBlock := dcrutil.NewBlock(b.chainParams.GenesisBlock)
 	header := &genesisBlock.MsgBlock().Header
-	node := newBlockNode(header, &stake.SpentTicketsInBlock{})
+	node := newBlockNode(header, nil)
 	node.inMainChain = true
 	b.bestNode = node
 
 	// Add the new node to the index which is used for faster lookups.
-	b.index[node.hash] = node
+	b.index.AddNode(node)
 
 	// Initialize the state related to the best block.  Since it is the
 	// genesis block, use its timestamp for the median time.
@@ -1455,12 +1455,10 @@ func (b *BlockChain) initChainState() error {
 		b.bestNode = node
 
 		// Add the new node to the indices for faster lookups.
-		prevHash := &node.parentHash
-		b.index[node.hash] = node
-		b.depNodes[*prevHash] = append(b.depNodes[*prevHash], node)
+		b.index.AddNode(node)
 
 		// Calculate the median time for the block.
-		medianTime, err := b.calcPastMedianTime(node)
+		medianTime, err := b.index.CalcPastMedianTime(node)
 		if err != nil {
 			return err
 		}
